Fix max int

+++ includes/color.h
@@ -6,7 +6,7 @@
 /*   By: alopez-g <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/09 03:52:34 by alopez-g          #+#    #+#             */
-/*   Updated: 2022/07/11 17:15:07 by alopez-g         ###   ########.fr       */
+/*   Updated: 2022/08/03 00:23:41 by alopez-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,6 @@
 
 void	mapper(t_stack st, int n);
 void	lstorder(t_stack *stack);
-t_list	*ft_lstmax(t_list *lst, int limit);
+t_list	*ft_lstmax(t_list *lst, long limit);
 
 #endif
